### 2020-07-23 #391 @DillonHammill
DataEditR: An Interactive R Package for Viewing, Entering & Editing Data
Submitting Author: @DillonHammill
Repository:   https://github.com/DillonHammill/DataEditR
Version submitted:   0.0.5
Version accepted: TBD

---

```
Package: DataEditR
Title: Interactive Data Viewing, Entry and Editing
Version: 0.0.1
    person(given = "Dillon",
           family = "Hammill",
           email = "Dillon.Hammill@anu.edu.au",
Description: Provides an interactive interface to view, enter or edit data
  or tabular data files in R.
  R(>= 3.5.0)
  shiny,
  shinythemes,
  rhandsontable,
  htmltools
Language: en-GB
```

`DataEditR` provides an interactive interface built on `rhandsontable` that supports viewing, entering and editing of data in R without the need for any coding. `DataEditR` allows users to interactively construct data.frames from scratch and edit/view existing data objects (e.g. matrices, data.frames and data.tables) or data saved to file (e.g. csv - any tabular data format is supported).

`DataEditR` has been designed to be as simple as possible to appeal to users with limited coding experience. It provides a familiar interface for viewing, entering and editing data that users of GUI-oriented software will be familiar with. `DataEditR` is also extremely useful for developers seeking an interactive interface to obtain inputs from the user. This package therefore has broad applications for any discipline that requires manual data entry and editing.

The `edit()` function from the `utils` package also offers the ability to interactively enter/edit data but lacks a lot of the features that one would expect from a GUI-oriented data manipulation tool (e.g. drag to fill, addition of checkboxes and dropdown menus). Not to mention the UI is not very visually appealing and there are undesirable glitches when entering data.

Perhaps the most similar packages to `DataEditR` that offer similar features are `DT` and `rhandsontable` (on which `DataEditR` is built). To the best of my knowledge, `DT` lacks the ability to interactively add/remove columns and rows and does not support editing of row/column names. For the most part, these packages were designed to interactively visualize datasets, whilst the focus the `DataEditR` is data entry and editing. `rhandsontable` has a lot of support for more complex visualisations of data, but the inner workings are quite complex (transferring data between R and javascript means that there are specific requirements for certain features to work). Currently, `rhandsontable` does not support editing of column names, this has been implemented in `DataEditR` through addition of custom javascript code. `DataEditR` has been developed over time to retain as much of the features of `rhandsontable` as possible whilst taking care of all the required data formatting internally.

The integration of reading/writing capabilities within `DataEditR` means that all the data loading, editing and saving operations can be performed in a single line (to minimize coding). To the best of my knowledge, `DataEditR` is the only R package that lets users interactively create data.frames from scratch without any coding.
