### 2020-07-27 053e0b0
Handle NA columns for dropdowns specified in col_options.
