### 2020-07-30 c6b1aa9
Changes for CRAN submission.
