+++ R/data_edit.R
@@ -21,15 +21,21 @@
 # may be possible to handle this externally but this gets complicated when
 # columns are added or removed.
 
-#' Interactively edit data.frames, matrices or tabular data files
+#' An interactive editor for viewing, entering & editing data
+#'
+#' \code{data_edit} is a shiny application built on \code{rhandsontable} that is
+#' designed to make it easy to interactively view, enter or edit data without
+#' any coding. \code{data_edit} is also a wrapper for any reading or writing
+#' function to make it easy to interactively update data saved to file.
 #'
 #' @param x a matrix, data.frame, data.table or the name of a csv file to
 #'   edit.Tibble are also supported but will be coerced to data.frames. An empty
 #'   table can be created by specifying the dimensions in a vector of the form
 #'   \code{c(nrow, ncol)}.
 #' @param col_bind additional columns to add to the data prior to loading into
-#'   editor, can be either an array containing the new data or a vector
-#'   containing the new column names for empty columns.
+#'   editor, can be either an array containing the new data, a vector containing
+#'   the new column names for empty columns or a named list containing a vector
+#'   for each new column.
 #' @param col_edit logical indicating whether columns can be added or removed,
 #'   set to TRUE by default.
 #' @param col_options named list containing the options for columns that use
@@ -39,9 +45,12 @@
 #' @param col_factor logical indicating whether character columns should be
 #'   converted to factors prior to returning the edited data, set to FALSE by
 #'   default.
+#' @param col_names logical indicating whether column names can be edited, set
+#'   to TRUE by default.
 #' @param row_bind additional rows to add to the data prior to loading into
-#'   editor, can be either an array containing the new data or a vector
-#'   containing the new row names for empty rows.
+#'   editor, can be either an array containing the new data, a vector containing
+#'   the new row names for empty rows or a named list containing a vector for
+#'   each new column.
 #' @param row_edit logical indicating whether rows can be added or removed, set
 #'   to TRUE by default.
 #' @param save_as name of a csv file to which the edited data should be saved.
@@ -84,16 +93,15 @@
 #' @author Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
 #'
 #' @examples
-#' \dontrun{
-#' # Edit data.frame save to csv
-#' data_edit(mtcars,
-#'   save_as = "mtcars-update.csv"
-#' )
+#' if (interactive()) {
+#'   # Edit matrix & save to csv
+#'   data_edit(mtcars,
+#'     save_as = "mtcars-update.csv"
+#'   )
 #'
-#' # Edit csv file
-#' data_edit("mtcars-update.csv")
+#'   # Edit csv file
+#'   data_edit("mtcars-update.csv")
 #' }
-#'
 #' @export
 data_edit <- function(x,
                       col_bind = NULL,
@@ -101,6 +109,7 @@ data_edit <- function(x,
                       col_options = NULL,
                       col_stretch = FALSE,
                       col_factor = FALSE,
+                      col_names = TRUE,
                       row_bind = NULL,
                       row_edit = TRUE,
                       save_as = NULL,
@@ -119,10 +128,10 @@ data_edit <- function(x,
[…]
