+++ README.md
@@ -102,17 +102,17 @@ citation("DataEditR")
 #> 
 #> To cite package 'DataEditR' in publications use:
 #> 
-#>   Dillon Hammill (2020). DataEditR: Interactive Editor for Tabular
-#>   Data. R package version 0.0.1.
+#>   Dillon Hammill (2020). DataEditR: An Interactive Editor for Viewing,
+#>   Entering & Editing Data. R package version 0.0.2.
 #>   https://github.com/DillonHammill/DataEditR
 #> 
 #> A BibTeX entry for LaTeX users is
 #> 
 #>   @Manual{,
-#>     title = {DataEditR: Interactive Editor for Tabular Data},
+#>     title = {DataEditR: An Interactive Editor for Viewing, Entering & Editing Data},
 #>     author = {Dillon Hammill},
 #>     year = {2020},
-#>     note = {R package version 0.0.1},
+#>     note = {R package version 0.0.2},
 #>     url = {https://github.com/DillonHammill/DataEditR},
 #>   }
 ```
