+++ DESCRIPTION
@@ -1,15 +1,16 @@
 Package: DataEditR
-Title: Interactive Data Viewing, Entry and Editing
-Version: 0.0.1
-Date: 2020-07-21
+Title: An Interactive Editor for Viewing, Entering & Editing Data
+Version: 0.0.2
+Date: 2020-07-30
 Authors@R: 
     person(given = "Dillon",
            family = "Hammill",
            role = c("aut", "cre"),
            email = "Dillon.Hammill@anu.edu.au",
            comment = c(ORCID = "0000-0002-1407-7223"))
-Description: Provides an interactive interface to view, enter or edit data 
-  or tabular data files in R.
+Description: An interactive editor built on rhandsontable to allow the 
+  interactive viewing, entering and editing of data in R. 
+  <https://dillonhammill.github.io/DataEditR/>.
 URL: https://github.com/DillonHammill/DataEditR
 BugReports: https://github.com/DillonHammill/DataEditR/issues
 Depends:
