### 2020-07-30 96e6c60
Quote rhandsontable in DESCRIPTION.
