+++ man/data_edit.Rd
@@ -12,6 +12,7 @@ data_edit(
   col_stretch = FALSE,
   col_factor = FALSE,
   col_names = TRUE,
+  col_readonly = NULL,
   row_bind = NULL,
   row_edit = TRUE,
   save_as = NULL,
@@ -55,6 +56,10 @@ default.}
 \item{col_names}{logical indicating whether column names can be edited or a
 vector of column names that cannot be edited, set to TRUE by default to
 allow editing of column names.}
+
+\item{col_readonly}{names of columns that cannot be edited. Users will be
+able to edit values but these will be reverted to the original values.
+Column names for these column cannot be edited either.}
 
 \item{row_bind}{additional rows to add to the data prior to loading into
 editor, can be either an array containing the new data, a vector containing
