+++ NEWS.md
@@ -1,3 +1,8 @@
+# DataEditR 0.0.4
+
+* Add `col_readonly` argument to prevent users from editing values or column names of certain columns.
+* Prevent error when column is clicked but not edited.
+
 # DataEditR 0.0.3
 
 * Add `col_names` argument to allow control over which column names can be edited by the user.
