### 2020-08-01 fdd1790
Make sure col_readonly and col_names can be used together.
