### 2020-08-06 6cd9f22
Fix row indices when rows are added or removed. Improve handling of duplicate rownames.
