### 2020-08-06 e4fe275
DataEditR v0.0.5.
