+++ vignettes/DataEditR.Rmd
@@ -226,7 +226,7 @@ mtcars_new <- data_edit(mtcars,
 
 ### 6.1 Checkboxes
 
-Checkboxes can be used to obtain TRUE or FALSE value inputs from users. The resulting data will have a value of 1 where TRUE was selected and NA for all other values.
+Checkboxes can be used to obtain TRUE or FALSE value inputs from users. The resulting data will have a value of TRUE or NA based on user input.
 
 ```{r, eval = FALSE}
 data_edit(mtcars,
