### 2020-08-06 cc41e6b
Write to tempfile in examples.
