+++ man/data_edit.Rd
@@ -119,12 +119,13 @@ function to make it easy to interactively update data saved to file.
 \examples{
 if (interactive()) {
   # Edit matrix & save to csv
+  file <- tempfile("mtcars-new", fileext = ".csv")
   data_edit(mtcars,
-    save_as = "mtcars-update.csv"
+    save_as = file
   )
 
   # Edit csv file
-  data_edit("mtcars-update.csv")
+  data_edit(file)
 }
 }
 \author{
