### 2020-08-09 8d8e1a7
Fix typo in vignette.
