### 2020-08-12 #391 @geanders
Thanks so much for your submission, @DillonHammill! It looks like a very interesting package, and one that will be useful for many R users.

I have two follow up questions:

First, a question regarding if the package is in-scope for rOpenSci: The package certainly looks like it could fall within "data munging", which is one of our categories. However, we also now have a special consideration for packages that are interactive:

> "Packages that include interactive tools to facilitate researcher workflows (e.g., shiny apps) must have a mechanism to make the interactive workflow reproducible, such as code generation or a scriptable API." (from Section 5.2.2 of the [rOpenSci Packages book](https://devguide.ropensci.org/policies.html#aims-and-scope))

For the purposes of this package, I think this would mean generating a code script that replicates all the interactive actions that the user has taken in data munging through the interactive interface the package provides. From the package README file, I didn't see any indication that a script is generated that could be used to replicate the interactive actions taken by the user. Could you please let me know if I've missed anything there?

Second, we typically ask for at least 75% coverage of package functions by unit tests, or an explanation for why coverage is lower (see Section 1.9 of the [rOpenSci Packages book](https://devguide.ropensci.org/building.html#testing)). It looks like this package does not have a testing suite. If the package is in-scope, we would need you to add a justification for the lack of a testing suite before the review.

Thanks very much for any details you can provide on either of these points!
