### 2020-08-12 #391 @DillonHammill
Thanks for taking a look at DataEditR @geanders. I hadn't thought about generating code to mimic the interactive edits, but it is certainly possible.

I haven't included unit tests at the moment as it is only a single function that is widely used in CytoExploreR. This is where I would pick up on any problems. I am working on adding some shiny tests but from memory these are not included in the test coverage anyway? I will also need to re-structure the package in order to be able to write shinytests for this package.

I will work on these and get back to you ASAP.
