### 2020-08-25 91f9996
DataEditR v0.0.6.
