+++ R/utilities.R
@@ -10,3 +10,12 @@ java_script <- function(...) {
   x <- paste(x, collapse = '\n')
   structure(x, class = unique(c("JS_EVAL", oldClass(x))))
 }
+
+## FILE_EXT --------------------------------------------------------------------
+
+#' Get file extensions from file names
+#' @noRd
+file_ext <- function(x) {
+  pos <- regexpr("\\.([[:alnum:]]+)$", x)
+  ifelse(pos > -1L, substring(x, pos + 1L), "")
+}
