### 2020-09-27 DillonHammill/DataEditR#5 @owbezick
@DillonHammill, I second the above idea. The author of the [editData](https://rdrr.io/cran/editData/man/editData.html) package (which provides similar functions) provides a module based approach, which could prove useful for those wanting to implement your GUI within their own shiny applications.
