### 2020-09-27 DillonHammill/DataEditR#5 @owbezick
Further, I have forked a repo to try to add the module functions. Please do let me know if you care to help!
