### 2020-09-27 DillonHammill/DataEditR#5 @DillonHammill
@owbezick, I would be happy to help add support this. In fact, I have a version where I have separated out the shiny components, so it may be easier to modify this version. I will push this to the master branch later today. I will also take a look at `editData` to see how they do things.
