### 2020-09-29 DillonHammill/DataEditR#5 @DillonHammill
@owbezick, I figured I would add some new features while making these changes, I should have a working version ready soon for you to test. I will let you know when it is ready.
