### 2020-10-02 DillonHammill/DataEditR#5 @leungi
Thanks for project @DillonHammill 🙏

This is a great light alternative to [{DTedit}](https://github.com/jbryer/DTedit) that I've used regularly, in case you'd like to avoid converting everything over to shiny modules for the short-term, but still make it compatible with shiny.
