### 2020-10-05 DillonHammill/DataEditR#5 @DillonHammill
Sorry @owbezick, I was really busy with other things last week. I should have more time to work on this now. I will try have it ready by the end of the week.
