### 2020-10-07 #391 @DillonHammill
Hi @melvidoni, I have had a few requests to convert the code to use shiny modules (https://github.com/DillonHammill/DataEditR/issues/5) so that the package can be used within other shiny applications and so I am currently working on this.

I have added a testing suite to the package but this will need to be updated once I finish making the shiny module changes.

In terms of returning the code to mimic the interactive edits made to the data, I think it is possible and I am exploring a number of options:

- Simply return the code required to generate the final edited version of the data. This is the simplest approach but does not display the steps taken to arrive at the final edited version.
- Record all data edits to a temp file that can then be converted into code once the application is closed. I don't think that this is the way to go, as users tend to get carried away when things are interactive and many of the changes made need not be reflected in the code output. There are some complex challenges to get this to work and this is why rhandsontable uses the approach above to output the code.
- Another alternative is that I simply return the code required to convert the starting data to the returned data. This is much simpler as I would only need to compare the initial and final `data.frames` and convert changes to code.

I am exploring ways in which I can use a combination of options 2 & 3 to generate the code in an efficient manner. It will take some time for me to properly implement this, so I am happy to resubmit or put this on hold until I have figured this out.
